Speed up display printing
use a 128 bit copy instead of regular copy_nonoverlapping. This requires
the display framebuffer to be aligned to 128 bits, but makes the update
of the display way faster.

Until there is some control of the MMU and the caches are enabled this
is pretty much what we can do in this domain, but should at least speed
up the development process.