Reduce bloat in LevelTable
Introducing another u8 in the LevelTable struct causes it to double the
allocation size (due to the alignment requirement to a page size).

Instead, we can simply keep track of the individual levels when
traversing the tree, saving half of the memory.

Reduce also the size of the early memory allocator to 64 kB.