Use fine-grained mappings for KLA
Kernel logical addresses are currently mapping the whole ram. This is
not ideal because everything has the same RWX permissions and memory
regions that should not be writeable are actually writeable.

Therefore, this commit changes it to create mappings for each specific
section with the proper permissions and also map the display memory
explicitly.