Remove spin crate dependency
We only need this for the mutex (which would be better called spinlock)
and can implement it ourselves (+ also implement with critical sections
as well if needed).

This change just replaces the crate, but does not add any new
functionality, a separate change will add the critical section with
support for nesting (e.g.: due to multiple spinlocks held at the same
time).