SpinLock should use critical sections
This would ensure that deadlock cannot happen due to a spinlock being
held while the exception is taken and later tried to be taken in the
interrupt context, which would deadlock.

In reality this is only a problem in the same core, so spinlocks across
different cores would still work fine.

For that to work we need to implement a per-cpu critical nesting
variable, but since there is no smp support so far this should be fine.