Add binary build for macOS 12.2 and onwards
macOS 12.2 onwards allows to load a custom boot object in binary format,
without having to use the Mach-O file format, which seems to be changing
behind the scenes. In order to take advantage of this feature and
install the OS persistently the new target will be needed.

Build a binary target with:

`cargo xtask build --release --binary`