Rename Uart to EarlyUart
The last remaining piece to remove the identity-mapping for mmio memory
is to properly map the UART. This commit renames the current uart work
as an early-uart version that we will use during boot (since we still
want this peripheral to work early in the boot chain).

After boot we will transition to a different UART implementation,
properly probed like the rest of devices that .

When the identity-mapping is removed TTBL0_EL1 will be able to be
switched to process mappings.