Migrate UART to use the print module
This creates a separate UART instance after relocation that already maps
via the memory manager the mmio region it needs instead of relying on
the identity-map.

In addition, it will be possible to extend it safely to have better
performance given that it is constructed after relocations are done.

Some logs are currently lost between the moment the kernel is relocated
and the moment we can initialize the late uart. This will be fixed later
by adding buffering.