Remove overlapping physical memory ranges
Even though the ADT gives us the information of the complete RAM
range, we may not use much of that due to carveouts. Some carveouts are
already accounted for, but unfortunately looks like there is more than
initially anticipated.

This change ensures that the available physical memory is only that
given to us by iBoot (or forwarded by m1n1) when we boot.